Moved Settings to Dedicated Files
