+++ src/components/ArcanaSettingsTab.ts
@@ -1,15 +1,20 @@
 import { App, Plugin, PluginSettingTab } from 'obsidian';
 import SettingsSection from './SettingsSection';
+import {
+  AvailablePlugins,
+  AvailablePluginSettings,
+} from '@/plugins/AllPlugins';
+import { AgentSettings } from '@/include/ArcanaSettings';
+
+export type AnyArcanaSettingSections = SettingsSection<
+  AvailablePluginSettings[AvailablePlugins] | AgentSettings
+>;
 
 export default class ArcanaSettingsTab extends PluginSettingTab {
   plugin: Plugin;
-  sections: SettingsSection<any>[];
+  sections: AnyArcanaSettingSections[];
 
-  constructor(
-    app: App,
-    plugin: Plugin,
-    sections: SettingsSection<any>[] // TODO: Type this more narrowly
-  ) {
+  constructor(app: App, plugin: Plugin, sections: AnyArcanaSettingSections[]) {
     super(app, plugin);
     this.plugin = plugin;
     this.sections = sections;
