+++ src/plugins/Nostradamus/NostradamusSettings.ts
@@ -1,0 +1,4 @@
+import { NostradamusSettings } from './NostradamusSettings';
+
+export type NostradamusSettings = Record<string, never>;
+export const defaultNostradamusSettings: NostradamusSettings = {};
